Actor: re-implement haste APR bonus
It used to be set in the effect queue, but that doesn't really work as,
for one, the fixed APR of most launchers would override it (items come
after spells following a reload).  Instead we're storing haste type in
the IE_IMPROVEDHASTE stat, and then there's a block late in actor stat
refresh code that manually adjusts APR based on that stat.

Further, type 0 and type 2 haste are supposed to round the APR to
integers, which they didn't before.  And, by being applied manually and
last, haste can now exceed the limit of 5 APR, which it should.