Update GUIOPT.py
Added gamecheck for IWD1 to the autopause options as per inline comment. The options for "enemy sighted", "trap found", "spell cast", and "center" are present in IWD1 but do not work correctly without the added check.  I have not yet tested the fix, but the option highlight does not appear on the options prior to adding the fix, after adding the fix the option highlights appear as for the other options.