View: refactor background drawing
we don't need to track dirty rects, especially because they could fall out-of-sync with reality (eg if a subview moved or resized)
Instead, draw background segments in DrawSubviews if needed.

This also means that MarkDirty() is no longer recursive, nor do we need a NeedsDrawRecursive(). Instead, just invalidate all subviews whenever a View is drawn.

this does a bit to facilitate partial redraws too, if we decide they would be worth it

overall this feels much more conceptually simple, and likely just as performant (I didnt measure). We don't have deep hierarchies and GameControl drawing will always be the most expensive, which we could now address better with partial redraws.

fixes #1340