GUIAnimation: remove SameResource()
seems like it was just a hack to work around GUIScript updates reassigning the same animation.

this may still be happening in some scripts somewhere, which now would restart the animation. I would wager that we can address that better on the script side. These days we have better tools at our disposal.