View: once again refactor background drawing
a few issues:
1. opacity and visibility are actually distinct
2. calling MarkDirty during drawing on non opaque subviews is problematic (return of #1340)

recursive background drawing was always a bit of a hack anyway, since it disregarded possible view contents aside from the BG. We had thought this wouldn't happen in practice due to the nature of the originals, but this was the cause of #1873, and GemRB does create and should support a richer hierarchy.

Instead, prior to determining what needs to be drawn for a given view, probe the subtree for subviews that would care about their superviews being redrawn (translucent and needing update) and mark those areas dirty so that everything that needs to be drawn is drawn during that views Draw() call.