Add flag to only highlight outline of doors and containers
Add the 'HighlightOutlineOnly'-flag, which toggles whether the game highlights
only the outline of doors and containers (like in BG1), or highlights the complete
surface of the object (all except BG1). In accordance with the original games,
this flag is enabled for BG1 and disabled for the other games by default.