BG1: fix radiobutton preselections in GUICG with associated cleanups
The radiobuttons in the GUICG windows suffer from the known radiobutton preselection
issue, i.e. the last created button of a buttongroup is preselected, which is not
the behaviour in the original and also leaves the controls in a wrong state, e.g.
the Done-button isn't enabled despite a preselected radiobutton.

Steps to fix this issue:
- call SetState(enabled/disabled) after SetVarAssoc in order to reset the selected state
- call SetVar(DictVar) after radiobuttons are created in order to restore the value of
  DictVar which was overriden by SetVarAssoc

While applying the above steps, the surrounding code contexts need cleanups and
simplifications:
- wrap male/female button into a function in GUICG1
- wrap Kit and KitName into functions in GUICG22 in order to improve readability
  of the for-loop
- group button settings (SetVarAssoc, SetFlags, SetState, ...) together and use
  the same sequence of settings for more clarity and consistency
- merge multiple for-loops over the same buttons into one for-loop
- simplify index and associated ControlID calculations
- add more comments to improve reasoning