GetOrient: replaced with a general implementation, trying to maintain speed
the way the old one did normalization ensured none of the odd orientations
could be returned. But we always want the correct orientation; if the
underlying animation does not have cycles for all of them, we already have
clamping in place to deal with that.

makes the test pass and improves projectile orientation, which is noticeable
for long ones, especially the scorchers when moving. Unfortunately I don't
see full effects on movement, where we sometimes still seem to be using only
8 orients (or the cycles are that similar).

based on:
https://stackoverflow.com/a/14100975/1494504