Button: make also the background text color configurable
works for bam fonts with the BACKGRND value set

doesn't work in the demo, since we don't pass this background flag to ttf
fonts. If you do that, you hit the odd composing we do in Font, so the
bg color usually outmatches the fg. If you skip that for non-default colors,
so only the fg is used, it stops affecting everyone, but you can still see a
slight border on bigger font sizes. So not much to gain.