Pathfinder improvements: traversability cache.
Implemented TraversabilityCache for caching positions of all Actors on the navmap.
The reasoning behind this is to unburden the execution hotpath in FindPath method, so that it won't call `Actor* Map::GetActor(const Point& p, int flags, const Movable* checker) const`, which is not cheap operation. Instead just make one memory access to the properly maintained cache, which stores information about each navmap cell: which Actor (if any) occupies the cell and whether it's bumpable or not.

The layout of data in the TraversabilityCache class and all its components was carefully chosen to be friendly to the data prefetcher and CPU data cache, validated by benchmarks on three different test benches representing different class of hardware:
- desktop with AMD Ryzen 9 5950x, fairly modern (2020), highend desktop chip, x86-64, 16C 32T, boost up to 4.9 GHz, cache L1 64 KB per core, L2 512 KB per core, L3 64 MB.
- laptop Dell Latitude E7240 with Intel Core i5-4300U, older mobile chip (2013), x86-64, 2C 4T, boost up to 2.9 GHz, cache: L1 64 KB per core, L2 256 KB per core, L3 3 MB.
- Raspberry Pi 4B Rev 1.2 with Broadcom BCM2711 SoC (2019), Quad core Cortex-A72 (ARM v8) 64-bit, 4C 4T, boost up to 1.5 GHz, cache: L1 32 KB data + 48 KB instruction per core, L2 1 MB.

The cache is updated once a frame, if and only if anyone called FindPath - there is no point in updating the cache if nobody wants to use it.
Updating once a frame means we're introducing a slight change of actors' perception when comes to finding path: previously they had immediate access to the newest actors' positions and if any changes in positions were done in the current frame, all the other actors knew about it right away. Now this information propagation is one frame delayed. There is no functional difference in behavior, bit there is a slight technical difference: this results in rare need of recalculation path without blocking actors where previously it wasn't needed, but also vice versa: no need of recalculation, where previously it was required. It's visible sometimes when measuring against original FindPath performance: speedup from 15ms to 200µs or (slowdown the other way around) means we encountered scenario where this difference manifests.

Additionally made some minor improvements in the FindPath method:
1) Removed check for the usage of the LazyThetaStar - the less branches in the hotpath, the better.
2) Made most data on which the algorithm works static - no need to make new allocations each call.

Those changes should yield improvements to the FindPath execution time across the board, but should be most noticeable for short paths: for the long and medium paths it's typically 1.2 - 1.5x faster, for short paths it's typically 1.3 - 2x faster. For extremely short paths it can vary: it could be 2x slower due to needed cache updates, but it could also be 3 - 4x faster; overall net positive. Detailed benchmark reports from multiple test cases and multiple machines are attached to the pull request.