Change handling of the cached navmap cells' state.
Change how we interpret the cells' state: don't treat is as arbitrary
enum values, rather as a token container.
There can be more than one actor occupying a cell, it must be able
to represent more than one traversability value at a time.
On entering a cell, each actor will now insert its token - increase the cell's
numerical value by (1) if being bumpable and (15) if not being bumpable.
The same value is subtracted when leaving a cell (token removal).

If a cell has value of 0, it means no actor is occupying it.
If a cell has value between 1 and 14, it means there are only bumpable
actors.
If a cell enters the treshold of 15, it means there is at least one
unbumpable actor.
This gives us ability to store 14 bumpable actors and 16
non-bumpable actors in one uint8_t cell.
Treating it as a tokens also removes the need of re-drawing neighbouring
actors, as now we don't blindly erase the cells' state, but maintaining it in a
more meaningful fashion.

Additionally introduce caching the blocking shape: there is no need to
recalculate blocking shape for each actor every cache update.