Pathfinder improvements: remove FibonacciHeap.
Removed FibonacciHeap used for storing and sorting nodes in the 'open' set.
Store this data in std::vector and use simple binary search to keep it sorted - this cuts off allocations and minimizes branch misprediction penalties.
Additionally removed one branch from the hotpath - choosing the function to call for getting child block status, based on size, which is constant through all the iterations.
Sprinkled the current code of `FindPath` with some `const` specifiers in places.